python-demo/requirements.txt
gbl-ninja-python